combine context file
