+++ src/contexts/UserPageContext.jsx
@@ -1,16 +1,22 @@
 import { createContext, useContext, useState } from "react";
 
-const defaultItineraries = {}
+const defaultItineraries = []
 
 const ItinerariesContext = createContext(defaultItineraries)
 export const useItineraries = () => useContext(ItinerariesContext)
 
 export const ItinerariesProvider = ({children}) => {
-  const [itineraries, setItineraries] = useState(defaultItineraries)
+  // provide user owned itineraries
+  const [itineraries, setItineraries] = useState([])
+
+  // provide user joined itineraries id
+  const [joinedItineraries, setJoinedItineraries] = useState([])
 
   return <ItinerariesContext.Provider value={{
     itineraries: itineraries,
-    setItineraries: setItineraries
+    setItineraries: setItineraries,
+    joinedItineraries: joinedItineraries,
+    setJoinedItineraries: setJoinedItineraries
   }} >
     {children}
   </ItinerariesContext.Provider>
